Move annotation filter chip into panel header
Inline the collapsed "All N · Filter" pill in the Annotations header row
instead of rendering it as a separate strip below. FilterBar now only
renders the expanded filter row; the filterLabel derivation moves to
SidePanel alongside the chip button. The filter-bar-toggle testid is
preserved so E2E helpers continue to work.
