+++ src/client/panels/FilterBar.svelte
@@ -11,7 +11,6 @@ interface Props {
   filterStatus: FilterStatus;
   hasFilters: boolean;
   open: boolean;
-  totalCount: number;
   onToggleOpen: () => void;
   onSetFilterType: (v: FilterType) => void;
   onSetFilterAuthor: (v: FilterAuthor) => void;
@@ -25,63 +24,15 @@ let {
   filterStatus,
   hasFilters,
   open,
-  totalCount,
   onToggleOpen,
   onSetFilterType,
   onSetFilterAuthor,
   onSetFilterStatus,
   onClearFilters,
 }: Props = $props();
-
-const filterLabel = $derived.by(() => {
-  if (!hasFilters) return "All";
-  const parts: string[] = [];
-  if (filterType !== "all") {
-    const labels: Record<FilterType, string> = {
-      all: "All",
-      highlight: "Highlights",
-      comment: "Comments",
-      note: "Notes",
-      "with-replacement": "With replacement",
-    };
-    parts.push(labels[filterType]);
-  }
-  if (filterAuthor !== "all") {
-    const labels: Record<FilterAuthor, string> = {
-      all: "Anyone",
-      claude: "Claude",
-      user: "You",
-      import: "Imported",
-    };
-    parts.push(labels[filterAuthor]);
-  }
-  if (filterStatus !== "all") {
-    const labels: Record<FilterStatus, string> = {
-      all: "Any status",
-      pending: "Pending",
-      accepted: "Accepted",
-      dismissed: "Dismissed",
-    };
-    parts.push(labels[filterStatus]);
-  }
-  return parts.join(" · ") || "All";
-});
 </script>
 
-{#if !open}
-  <!-- Collapsed summary chip -->
-  <div style="padding: 6px 16px; border-bottom: 1px solid var(--tandem-border);">
-    <button
-      data-testid="filter-bar-toggle"
-      onclick={onToggleOpen}
-      style="display: flex; align-items: center; gap: 4px; background: none; border: 1px solid var(--tandem-border); border-radius: var(--tandem-r-pill); padding: 3px 10px; font-size: var(--tandem-text-xs); color: var(--tandem-fg-subtle); cursor: pointer; white-space: nowrap;"
-    >
-      <span>{filterLabel} {totalCount}</span>
-      <span style="font-size: 9px; opacity: 0.7;">▾</span>
-      <span>Filter</span>
-    </button>
-  </div>
-{:else}
+{#if open}
   <!-- Expanded full filter row -->
   <div
     style="padding: 8px 16px; border-bottom: 1px solid var(--tandem-border); display: flex; gap: 4px; flex-wrap: wrap; align-items: center;"
